Move ncbi under taxa
Mirror the ontology heirarchy
